change file names and mv MultiModelError definition to BayesianInference
