--- usecases/linearFunction/Correlation_fine.yaml
@@ -1,16 +1,0 @@
-n_function_sensors: 50
-n_derivative_sensors: 0
-n_configurations: 3
-
-correlation_length_function: 0.3
-correlation_length_derivative: 0.3
-
-virtual_experiment_correlation_length_function: 0.3
-virtual_experiment_correlation_length_derivative: 0.3
-
-virtual_experiment_noise_std_function: 0.3
-virtual_experiment_noise_std_derivative: 0.3
-
-virtual_experiment_b: 3
-virtual_experiment_c: 2.5
-
